+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Hide "GIES Student" role button from onboarding and settings [#1121](https://github.com/rokwire/illinois-app/issues/1121).
+- Hide service initialization status in release builds [#1123](https://github.com/rokwire/illinois-app/issues/1123).
 
 ## [3.0.65] - 2021-12-10
 ### Added